[#71] feat: 목록 페이지 UI 변경

+++ real-estate-domain/src/main/java/io/gig/realestate/domain/realestate/construct/dto/ConstructCreateForm.java
@@ -34,7 +34,7 @@ public class ConstructCreateForm {
     private int useAprDay;
 
     // 대지면적
-    private int platArea;
+    private double platArea;
 
     // 건물면적
     private Double archArea;
