+++ real-estate-domain/src/main/java/io/gig/realestate/domain/realestate/construct/dto/ConstructDto.java
@@ -23,7 +23,7 @@ public class ConstructDto {
 
     private int useAprDay;
 
-    private int platArea;
+    private double platArea;
 
     private Double archArea;
